backend/requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
SQLAlchemy==2.1.2
apscheduler==3.10.1
python-dotenv==1.0.0
imapclient==2.3.0
beautifulsoup4==4.12.2
requests==2.31.0
flask-mail==0.9.1
pandas==2.2.2
